requirements.txt
Django>4.2,<4.3
djangorestframework>=3.16.0,<3.17
psycopg2>=2.9.10,<2.10
drf-spectacular>=0.28.0,<0.29
celery>=5.5.0,<5.6
redis>=5.2.1,<5.3
pyotp>=2.9.0,<2.10